requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
appnope==0.1.2
    # via ipython
backcall==0.2.0
    # via ipython
certifi==2020.12.5
    # via requests
chardet==4.0.0
    # via requests
decorator==5.0.9
    # via
    #   ipdb
    #   ipython
idna==2.10
    # via requests
ipdb==0.13.8
    # via -r requirements.in
ipython-genutils==0.2.0
    # via traitlets
ipython==7.24.0
    # via
    #   -r requirements.in
    #   ipdb
jedi==0.18.0
    # via ipython
links-from-link-header==0.1.0
    # via -r requirements.in
matplotlib-inline==0.1.2
    # via ipython
parso==0.8.2
    # via jedi
pexpect==4.8.0
    # via ipython
pickleshare==0.7.5
    # via ipython
prompt-toolkit==3.0.18
    # via ipython
ptyprocess==0.7.0
    # via pexpect
pygments==2.9.0
    # via ipython
python-dotenv==0.17.1
    # via -r requirements.in
requests==2.25.1
    # via -r requirements.in
toml==0.10.2
    # via ipdb
traitlets==5.0.5
    # via
    #   ipython
    #   matplotlib-inline
urllib3==1.26.5
    # via requests
wcwidth==0.2.5
    # via prompt-toolkit

# The following packages are considered to be unsafe in a requirements file:
# setuptools
